migration refactor WIP
