+++ src-esp/main/rfid/mock_rfid.cpp
@@ -1,0 +1,10 @@
+#include "rfid.h"
+#include <string>
+#include <iostream>
+
+std::string test_str;
+void setup_rfid_reader()
+{
+    test_str = "Mock tester";
+    std::cout << "Running setup for " << test_str;
+}
