+++ src-esp/main/rfid/pn532.c
@@ -1,0 +1,8 @@
+#include "rfid.h"
+
+
+void setup_rfid_reader()
+{
+    //Make task for monitoring RFID
+
+}
