+++ src-esp/main/tusb_hid_example_main.c
@@ -7,17 +7,20 @@
 #include "constants.h"
 #include "cards.h"
 #include "comms.h"
+#include "main.h"
+#include "rfid/rfid.h"
+#include "util/profile.h"
 
 #include <inttypes.h>
 
 #include <stdlib.h>
-#include "esp_log.h"
+#include <esp_log.h>
 #include "freertos/FreeRTOS.h"
 #include "freertos/task.h"
 #include "tinyusb.h"
 
 #ifdef LOG_DETAILS
-#include "esp_chip_info.h"
+#include "util/profile.h"
 #endif
 
 #include "tusb_cdc_acm.h"
@@ -30,8 +33,8 @@
 #include "nvs_flash.h"
 #include "nvs.h"
 #include "esp_timer.h"
-#include "rc522.h"
 #include "cJSON.h"
+
 // https://wokwi.com/projects/395704595737846785
 
 /********* Application ***************/
@@ -137,19 +140,6 @@ static void app_send_hid_demo(void)
 }
 
 /*******KEYDOTBOARD APP ******/
-typedef enum
-{
-    APP_STATE_BOOT,
-    APP_STATE_SCANNER_MODE,
-    APP_STATE_MASTER_MODE,
-    APP_STATE_APPLY_KEYSTROKES,
-    APP_STATE_SEND_PASSWORD_DB,
-    APP_STATE_SEND_RFID,
-    APP_STATE_SAVE_NEW_CARD,
-    APP_STATE_CLEAR_CARD,
-    APP_STATE_CLEAR_DB,
-    APP_STATE_MAX
-} APP_STATE;
 
 typedef enum
 {
@@ -187,12 +177,11 @@ static const char *RESPONSE_TYPE_STR[RESPONSE_TYPE_MAX] =
 
 APP_STATE state = APP_STATE_BOOT;
 nvs_handle_t storage_handle;
-rc522_handle_t scanner;
 RFID_DB_t rfid_db;
 
 uint64_t currently_scanned_tag = 0;
 NEW_CARD_t current_new_card;
-char currently_scanned_pass[MAX_PASS_SIZE];
+int currently_scanned_tag_index = -1;
 int current_clear_card_index;
 
 void get_pass_from_id(size_t in_selected_id, char *out_pass)
@@ -217,51 +206,6 @@ void get_pass_from_id(size_t in_selected_id, char *out_pass)
     // TODO return esp error value or make own enum for error handling
 }
 
-static void rc522_handler(void *arg, esp_event_base_t base, int32_t event_id, void *event_data)
-{
-    rc522_event_data_t *data = (rc522_event_data_t *)event_data;
-
-    switch (event_id)
-    {
-    case RC522_EVENT_TAG_SCANNED:
-    {
-        rc522_tag_t *tag = (rc522_tag_t *)data->ptr;
-        ESP_LOGI(TAG, "Tag scanned (sn: %" PRIu64 ")", tag->serial_number);
-
-        switch (state)
-        {
-        case APP_STATE_SCANNER_MODE:
-        {
-            currently_scanned_tag = tag->serial_number;
-            state = APP_STATE_SEND_RFID;
-            break;
-        }
-        case APP_STATE_MASTER_MODE:
-        {
-            // Check if the tag is in the database
-            for (size_t i = 0; i < rfid_db.total_rfid_tags; i++)
-            {
-                if (rfid_db.serial_number_buffer[i] == tag->serial_number)
-                {
-                    char outpass[MAX_PASS_SIZE];
-                    get_pass_from_id(i, outpass);
-                    strcpy(currently_scanned_pass, outpass);
-                    // TODO: remove pass logging
-                    ESP_LOGI(TAG, "Found tag %" PRIu64 " in db, pass: %s", rfid_db.serial_number_buffer[i], outpass);
-                    state = APP_STATE_APPLY_KEYSTROKES;
-                    return;
-                }
-            }
-            ESP_LOGE(TAG, "Couldnt find Found tag %" PRIu64 " in db", tag->serial_number);
-        }
-
-        default:
-            break;
-        }
-    }
-    break;
-    }
-}
 // https://docs.espressif.com/projects/esp-idf/en/stable/esp32/api-guides/performance/size.html#idf-py-size
 // This also helps https://docs.espressif.com/projects/esp-idf/en/stable/esp32/api-reference/system/mem_alloc.html#_CPPv425heap_caps_print_heap_info8uint32_t
 
@@ -429,6 +373,9 @@ void wait_for_hid_ready()
 void send_password_keystrokes()
 {
 
+    char currently_scanned_pass[MAX_PASS_SIZE];
+    get_pass_from_id(currently_scanned_tag_index, currently_scanned_pass);
+
     // release all keys between two characters; otherwise two identical
     // consecutive characters are treated as just one key press
 
@@ -480,66 +427,10 @@ void send_password_keystrokes()
     ESP_LOGI(TAG, "Password sent");
 }
 
-#ifdef LOG_DETAILS
-esp_chip_info_t chip_info;
-
-void print_memory_sizes(void)
-{
-
-    // uint32_t flash_size = ESP.getFlashChipSize();
-    // printf("--------> Flash size: %PRIu32 bytes\n", flash_size);
-
-    // Flash Size
-    const esp_partition_t *partition = esp_partition_find_first(ESP_PARTITION_TYPE_APP, ESP_PARTITION_SUBTYPE_ANY, NULL);
-    if (partition)
-    {
-        ESP_LOGI("Memory Info", "Found App partition");
-        ESP_LOGI("Memory Info", "Partition Label: %s", partition->label);
-        ESP_LOGI("Memory Info", "Partition Type: %d", partition->type);
-        ESP_LOGI("Memory Info", "Partition Subtype: %d", partition->subtype);
-        ESP_LOGI("Memory Info", "Partition Size: %" PRIu32 " bytes", partition->size);
-    }
-    else
-    {
-        ESP_LOGE("Memory Info", "Failed to get the App partition");
-    }
-
-    // Total SPIRAM (PSRAM) Size
-    size_t spiram_size = heap_caps_get_total_size(MALLOC_CAP_SPIRAM);
-    if (spiram_size)
-    {
-        ESP_LOGI("Memory Info", "PSRAM Size: %zu bytes", spiram_size);
-    }
-    else
-    {
-        ESP_LOGI("Memory Info", "No PSRAM detected");
-    }
-
-    uint32_t total_internal_memory = heap_caps_get_total_size(MALLOC_CAP_INTERNAL);
-    uint32_t free_internal_memory = heap_caps_get_free_size(MALLOC_CAP_INTERNAL);
-    uint32_t largest_contig_internal_block = heap_caps_get_largest_free_block(MALLOC_CAP_INTERNAL);
-
-    ESP_LOGI("Memory Info", "Total DRAM (internal memory): %" PRIu32 " bytes", total_internal_memory);
-    ESP_LOGI("Memory Info", "Free DRAM (internal memory): %" PRIu32 " bytes", free_internal_memory);
-    ESP_LOGI("Memory Info", "Largest free contiguous DRAM block: %" PRIu32 " bytes", largest_contig_internal_block);
-}
-
-void print_system_info_task(void *pvParameters)
-{
-    while (1)
-    {
-        size_t free_memory = esp_get_free_heap_size();
-        printf("Free Memory: %d bytes\n", free_memory);
-
-        vTaskDelay(pdMS_TO_TICKS(1000));
-    }
-}
-
 void print_state_cb(void *arg)
 {
     ESP_LOGI(TAG, "State: %d", state);
 }
-#endif
 
 void app_main(void)
 {
@@ -611,18 +502,7 @@ void app_main(void)
 
     init_rfid_tags();
 
-    rc522_config_t config = {
-        .spi.host = SPI3_HOST,
-        .spi.miso_gpio = GPIO_NUM_11,
-        .spi.mosi_gpio = GPIO_NUM_9,
-        .spi.sck_gpio = GPIO_NUM_7,
-        .spi.sda_gpio = GPIO_NUM_5,
-    };
-    // TODO: Add error handling
-    rc522_create(&config, &scanner);
-    rc522_register_events(scanner, RC522_EVENT_ANY, rc522_handler, NULL);
-    // Dont need to pause the scanner in whatever mode we operate under
-    rc522_start(scanner);
+    setup_rfid_reader();
 
     state = APP_STATE_MASTER_MODE;
 
@@ -652,6 +532,19 @@ void app_main(void)
                 app_send_hid_demo();
                 send_serial_msg();
             }
+
+            break;
+        }
+        case APP_STATE_SCANNED_CARD:
+        {
+            // Test if trigger is pressed
+
+            break;
+        }
+        case APP_STATE_TRIGGER_BUTTON_PRESSED:
+        {
+            // Test on release
+            // if detected, kill blink routine and change state
 
             break;
         }
